Fix issues with JSON nulls.
1. Track/Recording Lengths

Most of these used a `plain GetDouble()`; but recordings can have 'null'
as length, causing an exception.
Moved `HelperMethods` to the `Json` namespace (only place it's used from
anyway) and added a `GetOptionalTimeSpanFromMilliseconds()` extension
method for convenience when reading length fields.

2. Release Status Id

This used a plain `GetGuid()`; updated it to `GetOptionalGuid()`.

Fixes #3.