Switch `OAuth2` to use `HttpClient`
It is now `IDisposable`, to enable cleaning up the `HttpClient`.

It can be constructed with a client object to use. The ownership of that
client can stay with the caller, in which case that is the only client
the OAuth client will use, and `Close()` throws an exception.

Otherwise, an `HttpClient` instance will be created whenever needed, and
`Close()` can be used to ensure any such client is cleaned up (with
the next request creating a new one). Both the creation and further
configuration of such clients can be done via user code.

Either way, requests will always:
- use absolute URIs constructed using the `UrlScheme`, `Server` and
  `Port` properties (so any base address set on the client will be
  ignored)
- use `application/json` for the `Accept` header (so any default
  value configured on the client for that header is ignored)
- add the library name and version at the end of the `UserAgent` header
  (so any default value configured on the client for that header is
  honored)

This also adds a bunch of utility methods in a `Utils` class; in time
these may be moved to a `MetaBrainz.Common` package.