Update user info retrieval
The email address in the user info is now considered optional (because
it is only returned when the `email` scope has been granted). The
documentation for `GetUserInfo[Async]` now mentions this, as well as the
requirement for the `profile` scope to have been granted.

This also renames `AuthorizationScope.EMail` to
`AuthorizationScope.Email`.