Update `MetaBrainz.Build.Sdk` to version 4.0.0
This means we now target `net8.0` only.

This also tweaks the build, requesting GA .NET SDKs only (so no
previews), and passing `--skip-duplicate` on the NuGet push, to make it
easier to rerun the build for a tag if something goes wrong during
publishing.

Hexadecimal flag enum handling is enabled for the API reference, because
most of the MusicBrainz flag enums are indeed bitfields.