Rework internals
The existing assumption was that no endpoint expects a parameter to
appear more than once. But the URL lookup now allows multiple "resource"
values to be specified, so the internals need to change to accomodate.

This does mean that almost every request now requires additional
allocations.

Global usings are used to make it easier to change the data types used
for the query options with a minimum of source changes.